Build all of gfxspy.
Added implementations for all of OpenGLES and EGL API.
The API implementation is generated using the pre-processor
to expand macros found in trace.in and egl_trace.in.

libgfxspy.so can now be linked in as a direct replacement
for libGLESv2 and libEGL.

Modified the loader so that the hooks call the system versions
of GLES and EGL functions, instead of the driver versions.

Removed trailing white-space.

Change-Id: I6da4202f4e894dab6dce8e48365d26150dfe1468