Add support for more number kinds to Reader.Numeric
It now hex, octal, floating point and scientific representations.
Non-floating point numbers can have an optional "u" suffix.  Note that
the function now does not consider the leading sign to be a part of
the number. I think this is more correct, since in most cases the
Reader cannot determine whether the sign is a unary or a binary
operator and this would cause trouble for the parser.

Change-Id: I8dd460ccd871ebc9a2deda8ba654bd7dad4f13f6