Support depth buffer readback on GLES 3, keeping GL compatiblity.
Works in most cases, unless the depth attachment can't be matched with
a depth texture, e.g. GL_DEPTH_COMPONENT16/GL_DEPTH_COMPONENT at the
moment. This method depends on GLES 3 due to the glBlitFramebuffer call.

It may be possible, with a more intrusive approach, to support GLES 2 by
intercepting the last framebuffer depth attachment binding and replace it
with a custom depth texture but this could result in unwanted side effects
for the (potential) rest of the stream.

Change-Id: Ie6bafc3e1d7d93b6f5f982b8e61b8b1a39abf191