Remove count from atom stream.
The new graphics spy will be creating atoms streams and will not know
ahead of time how many atoms it is going to produce. Use the EOS atom to
indicate there are no more atoms across all contexts (which are soon to
change).

This is the first CL extracted from:
https://android-review.googlesource.com/#/c/143742

Change-Id: I3a59a3e4039568dcf5daf81eb2f7405a951fa3b7