Use the new spy inheritance to implement a proper init()
Get the window width, height and formats on wglCreateContext and emit an
init() atom storing this information.

Correctly re-resolve the imported functions on wglMakeCurrent. These could
be cached per-context, but I don't care too much about performance on
windows.

Update the encoder to match the latest changes to the atom stream format.

Fix bad Encoder ownership with a shared_ptr.

Change-Id: I634aee2e8deddc67fe236e313c9fec1186ef2394