Begin golang build scripts as a replacement for gradle.
I've lost too much time trying to get gradle to build our C++ projects.
I'm sure it's possible to get gradle to do what we need, but the lack of
documentation, stupidly slow build times and flakyness when off the
internet as made me give up on it for now.

Change-Id: I0d0379394624aed7028f6f7372d898557bbffccf