Template support improvements
Added an equivalent to Macro called Template that invokes the specified
template and writes it's output directly to the underlying stream rather
than returning it as a string.
Added a Node function that attempts to dispatch to a template with a
specified prefix that matches a node's information. This removes the
need for many of the cascading if's in templates, and makes them much
faster (and also more powerful)

Change-Id: Iac79fcda464341cfd32bf355d9f93f82997be4c5