Add support for casting char* -> string.
Attempting to do this with strlen and slicing does not work as we
need to observe the null-terminator, but don't want it in the
resulting string.

This CL allows casts from char* to string, which does the correct
observations and omits the null-terminator from the string.

This also allows us to kill the strlen extern.

Change-Id: Ia60a589340a6cc627c7d2ddc781e2d89a3f2f664