Fixup the atom class delegation so you can modify atoms again
The Atom wrapper was returning the schema.Class directly, which breaks
encoding because schema.Class does not know how to encode the atom
wrapper.
Now it returns the AtomClass wrapper as it's class, which knows to
extract the wrapped schema.Object and then use the schema.Class to
encode it.

Change-Id: Ief47c7629c29f9bf17f84d3758c48ecadd0af98d