Refactoring of the go client.
Instead of keeping all the requested in a big-fat application
context, keep the necessary information in each of the panels.
This:
 • Reduces chances of data-races with updates between panels.
 • Means we can have side-by-side views on different parts of
   the capture, or even different captures.
 • Let's each panel listen for paths they're interested in, or
   could let us support drag-drop paths from one panel to
   another.

We now also use the new task package for cancelable,
asynchronious jobs, like RPC calls.

RPCs and events have also been moved to separate structures to
make the application-context less monolithic.

The client behaves as it did before, except for pointers are
currently non-clickable. I'll be reimplementing this with paths
soon.

Change-Id: I2f1d89ca73534c68dc159882f2d7ce91ae870c49