Improve ClassName
It now digs into the resolved module for the actual struct, and uses the
tags on the struct to fix the typename itself as well.
This allows name remapping to work correctly even for fields of structs.
It also now handles types from the same module correctly.
This allows it to be used for all places were we emit a class name in
java.
Use the new features to rename atom.Group in to to AtomGroup in java
correctly to avoid ambiguity.

Change-Id: I58d89387b2cf2cdc1d29267692bdfaa6baa62a55