Have gapii/gapit select the ABI based on the package information.
We already call "dumpsys packages" which contains the information
about package ABI. I have added code to parse this. If the package
ABI is specified then that is used in preference to the device ABI.
This allows tracing 32-bit applications on 64-bit devices running M.
There is a bug in L which means wrapped 32-bit applications don't work.

b/18713965

Change-Id: I39df16fcae8767fdb1642650c87b352b4b2b2fcb