Partial file rewriting support
The template system now uses the supplied template name if the file does
not exist or contains no markers.
If it finds section markers of the form
<<<Start:templatename>>>
<<<End:templatename>>>
then the text between the markers is replaced by the result of
evaluating templatename

This allows the codergen to output files that can then be hand edited
and still be partially update by codergen.

Change-Id: Ibfc690ab85db920c81491ec794c8de4ad98c3237