Java client fixes
Make the main client an abstract class instead of an interface.
Lowercase all the method names to match java style.
Rename the impl to RPC
Add an impl that forwards to an underlying client, this allows
interception of some messages, will be used in a caching client.
Fix issues with duplicate imports.

Change-Id: I301342d3228a651efd6d6e5c35c876e050de4900