Fix assignment to temporary texture unit expressions in the API files.
We had a map of maps, and were attempting to mutate map keys that were
not in the outer map. This would cause default initialized maps to be
returned that were not automatically added to the outer map, resulting
in pointless mutations to a temporary, short-lived map.

Instead create a class for each TextureUnit that are constructed
up-front and placed into the map.

Change-Id: I8dc62eda5adc9a32f627b879f45fc1453f072a28