Remove Skip method from binary.Object and binary.Decorder.
Skipping was only used in a rare case where you rewind over
an object you already decoded. We now redecode the object
and throw away the result. Doing this saves hundreds of Kb
of generated code. Also move objects/nil.go to registry/nil.go
so that it isn't possible to accidentally fail to register it.

Change-Id: I4427a54b58b60644be95ff4c35f2fdb043a135c7