Provide a mechanism for reading old capture files.
Code generated structs maybe frozen to preserve the ability to decode them.
Encoding of the frozen structs is not possible. At the moment only
manual upgrading is provided (i.e. the programmer must provide an
upgrade method).

Note ID generation does not work yet. This is not new, but means that
this change alone is not sufficient to give us the ability to read
old streams.

As implemented here a change at the base of a containment heirarchy
will require versioning all the structs above it in the heirarchy.
This potentially leads to explosion of typeIDs and generated code.
Potentially O(n * c) where n is the number of non-frozen types and
c is the number of changes.

There are alternatives to this approach such as composing IDs or
using a schema wide revision number. These approaches lead to a more
palatable O(n + c) growth, but are hard to do without making changes
to the existing format. Hopefully this CL will make give us a basis
for discussing which direction to go.

Avoid version ripples in composition hierarchies.

This change is using the approach of encoding sub-type IDs in the
stream. At the moment no Java or C++ support. This is a bit "early
access", some of the generated code is a bit smelly. Need some more
functionality in codergen to do better, though.

Make the split between flat and cyclic more distinct.

Also change Multiplexer to use flat (in Go).

Added more defensive programming.
There is some printf debugging to be removed.
TypeID() and ID() are now distinct.
Generated init() is now a single one at the end of the file.
Need to do something for schema/subspace of frozen objects.

Change-Id: Ie58c1a8866c95d0d1cd5467711b39e091a16e1a4