Killing error returns on binary read/write calls
The mix of error returns and error on the stream was making life harder when writing lots of by hand encoder stuff for the schema, so I decided to fix it.
I kept the newly generated binary files in a separate cl for ease of review (all changes in this cl were made by hand), but it won't work without both.
There are some very minor changes in behaviour in the presence of errors, but the non error case should be identical.

Change-Id: I2b9b15983262ff05f5b81d8eafef0b67b503df2f