Types in the schema package are no longer processed by codergen
All encode/decode functions are now hand written, and don't involve the
type system.
This is in preparation for the schema becoming the type system.

Boxing is now sane. Boxing is the responsability of the binary package,
and supports pluggable boxers.
The Any schema type is part of the schema package.
The type specifix boxes are still in the any package, but are automatic.
This allows the any package to depend on the schema package, and thus
have a schema the way it should do, and removes the need for the
code/encode stuff to depend on the any package.

Change-Id: I85d1ab46f783675a7b02ef869ac38d4b26f58889