schema.Entity -> binary.Entity
schema.Type and schema.Field had to move as well as they are used in
Entity.
ObjectClass took over from Entity as the binary.Class for dealing with
dynamic objects.
This is to prepare for Entity appearing direclty in the binary
interfaces, so it has to be in the binary package.

All auto generated changes come in the next cl for easier reading.

Change-Id: I021458e1059cce16d5cc3e72999edd12d44fb12f