True schema graph encoding
This adds an entity specific encode/decode method, and a separate entity
table. Entities now encode the entire graph recursivly using this
method, so that the schema no longer needs type id's to rewire itself.
This also replaces the service.Schema method with a custom object
(schema.Message) that knows how to use this encoding system for the
schema objects.

Change-Id: I8ba5d9828353c9c8e6b1a62dd1a6027ad9f83e28