Add the concept of a control block to a binary stream
sid 0 encoded true was previously an invalid combination, as sid 0 is
the special nil sid and thus never encoded. This magic pair now
indicates a stream control block is present in the stream
The stream control block is versioned, and in the first version it only
supports an encoding mode, that switches schema between compact and full
forms.
Compact is the defualt. This means the existing .gfxtrace files are
still valid, as they will have no control blocks. It does change the
schema rpc message however.

Change-Id: Ie875250d2d86a03cf0f314d2a6314a76e65d0556