Fix for T** arguments when capture is 32bit and replay is 64bit.
Previously, the pointers would be patched to replay addresses in-place,
ignoring the fact there was not enough room to hold the pointer.

Funnily enough, this would cause crashes.

With this CL, all heap pointers are stored in a pointer table that is
separately managed from regular ObservedPointers.

This doesn’t solve other variable T types like int**, but I’ve yet to
encounter an API doing this.

Change-Id: I678c6148a012d3474009b672f36b4cb8e5b41367