Replace gapic::Array with gapic::ScratchAllocator and gapic::Vector.
Array was a union of std::Vector and a raw pointer / size pair.
This was a really horrible interface, and either meant you’d hit the 
heap with frequent small allocations, or have to jump through scary
hoops with object lifetimes.

Instead replace the use of gapic::Array with gapic::Vector. The new
vector type simply wraps a pre-allocated buffer, and guards against
buffer overrun / under-runs.

To simplify dealing with object lifetimes we also now have 
gapic::ScratchAllocator - a simple linear allocator that gets reset at
the end of each atom.

Paired together, we get a simple and efficient way to encode atoms and
their payloads.

Change-Id: I589320dbe33c7fbebeef45b6fe2d77a9204afccd